requirements.txt
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.2
Django==4.0.1
django-allauth==0.47.0
djangorestframework==3.13.1
drf-spectacular==0.21.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.4.0
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
